backupccl: add 'check_files' option to SHOW BACKUPS
Previously, the user did not have an easy way to check that all SSTs belonging
to a backup were in the expected place in external storage. This patch adds the
'check_files' option to SHOW BACKUPS which checks that all SSTs in a backup
chain are in their expected location in external storage. If SHOW BACKUPS cant
read from a file, an error message gets returned to the user about which file in
which directory was unable to get read.

Release justification: low risk, high benefit changes to existing functionality
Release note (sql change): adds the 'check_files' option to SHOW BACKUPS to
allow users to verify their SSTs are in their proper place.