importer: add ImportEpoch table descriptor field
This patch introduces the new ImportEpoch table descriptor field which
provides a count for the number of IMPORT INTO jobs that have been attempted on
the table, including any in-progress jobs. The epoch is incremented during the
prepareExistingTablesForIngestion call in the import job.

This work is apart of a larger project to roll back imports without MVCC
timestamps. In a future pr (#85138), this epoch will get written to the
imported MVCCValueHeader metadata, which in turn allows import rollbacks to
use the epoch to determine which keys to delete in the table.

We considered binding the import job id to the table descriptor, and in turn
to each imported KV's metadata, however, the expected proto encoded size of a
job ID was between 7 - 12 bytes, which seemed uncessary. Meanwhile the 4 byte
uint32 epoch will encode to 1 byte for the first 256 IMPORT INTO jobs on the
table. 2,147,483,648 IMPORT INTO jobs must occur on the table before we worry
about integer overflow -- a problem for a future release :D.

Informs #85138

Release note: none