storage: add importEpoch predicate to DeleteRange
This patch adds an ImportEpoch field to the predicate that clients can pass to
DeleteRange. When the client passes the ImportEpoch predicate, DeleteRange will
delete keys whose latest version contains the matching ImportEpoch in its
MVCCValueHeader.

The typical use case for predicate based DeleteRange will be rolling back and
IMPORT on a table with pre-existing data that was restored after the IMPORT
began. In this situation, the pre-Import and Imported data's timestamps are
intermingled; so during a rollback, the invariants assumed in
Predicate.StartTime DeleteRange do not hold. Instead, rollbacks must rely on
the ImportEpoch stored in each imported versioned key's MVCCValueHeader.

Since the cannot pass a Predicate.StartTime to ImportEpoch based
DeleteRange, time bound iteration is not used, implying
that ImportEpoch based DeleteRange can be significantly slower.

Informs #76722

Release note: none