backupccl: add tpcc cluster replication roachtest
This pr adds an end to end cluster replication roachtest described in #92697.
Specifically, the following occurs:

0. Initialize  source and destination clusters, each with 4 nodes, running
8vcpu machines.
1. Bulk init the tpcc workload with 500 warehouses, resulting in about 30 GB of
data. This takes around 10 minutes.
2. Start a replication stream and a tpcc workload set to run for
estimatedInitialScanTime + 10 minutes of steady state replication. The
estimatedInitialScanTime is set to the bulk import time, defined in step 1.
3. Cutover to the dest cluster at a timestamp that's 5 minutes before the
workload defined in 2 is expected to end. This takes less than a minute.
4. Fingerpint the src and dst clusters. Currently takes 15 mins total, with
archaic fingerprint query.

So, this test currently takes a whopping 45 minutes to complete. Once the new
fingerprinting job is used in this test, the runtime will reduce by  14
minutes. Longer term, we should init the src cluster with disk snapshots as
well. As this test becomes more efficient, we ought bump the number of
warehouses to 1000.

A follow up PR will also add infrastructure to send initial scan and cutover
time benchmarks to roachperf.

Informs #92697

Release Note: none