c2c: randomize node shutdown time in reslience roachtest
Previously, the c2c/nodeShutdown tests would only execute a shutdown after a
high water mark was set and before a cutover timestamp was chosen. This patch
randomizes the node shutdown to either occur during the initial scan, during
this steady state phase, or during cutover.

This patch also fixes a funky bug that caused a run of the c2c/shutdown tests
with the `--count` flag to produce total nonsense results because test state
leaked across count runs. Specifically, this patch ensures the
replicationTestSpec is instantiated when the test driver is called rather than
when the test is registered.

Informs #89487

Release note: none