c2c: use new in memory sst on range key batcher reset
Previously, the rangekey batcher would reuse the in-memory sst after flushing
an sst to the kv layer. Unfortunately, this can cause a subsequent rangekey
flush to corrupt the keys in the previous rangekey flush, as observed
 in #104203.

This patch fixes this bug by making the rangekey batcher look more like the
sstbatcher: during the post flush reset, an new memory sst is allocated.

Fixes #104203

Release Note: none