logic: skip_on_retry works when errors are expected
Previously, we have `skip_on_retry` directive for logic test which, when
set, it skips the rest of test if a statement fails with
TransactionRetryError. However, it won't skip if the statement is
expected to fail with certain error message. This PR ensures that
whenever we have a TransactionRetryError and `skip_on_retry` is set, we
always skip the rest of the test, even if the stmt is expected to fail.

Informs #104464

Release note: None