roachtest: add jobs contention roachtest
This patch adds a roachtest that attempts to stress the job system. To do so,
the test spins up 5,000 tables with changefeeds with a 5 minutes pts
expiration, randomly pauses and resume these changefeeds to stress the pts
management poller, and polls SHOW JOBS to assert that it returns in less than 2
minutes. If the assertion fails, the debug zip will hopefully reveal some form
of contention on the job system.  Further, the test also runs EXPLAIN ANALYZE
SHOW JOBS after the assertion fails for more debugging information.

Informs #118512

Release note: none